requirements.txt
# Framework Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
joblib==1.3.2

# Ollama (Python client)
requests==2.31.0

# Utilidades
python-dotenv==1.0.0
apscheduler==3.10.4

# Opcional: Mejoras
scipy==1.11.4
plotly==5.18.0